+++ include/CShmArray.h
@@ -79,7 +79,7 @@ namespace lce
 		//create shm
 		if ( m_oShm.getShmID() <= 0 || !m_oShm.attach() )
 		{
-			if ( !m_oShm.create(iShmKey, (int)dwShmMaxSize, bCreate, bReadOnly) )
+			if ( !m_oShm.create(iShmKey, dwShmMaxSize, bCreate, bReadOnly) )
 			{
 				snprintf(m_szErrMsg, sizeof(m_szErrMsg), "shm init create error: %s", m_oShm.getErrMsg());
 				return false;
